+++ krbticket/__init__.py
@@ -1,2 +1,3 @@
 from krbticket.ticket import *
+from krbticket.updater import *
 from krbticket.config import *
